Module Make.Storage

Tezos Protocol Implementation - Typed storage

This module hides the hierarchical (key x value) database under pre-allocated typed accessors for all persistent entities of the tezos context.

This interface enforces no invariant on the contents of the database. Its goal is to centralize all accessors in order to have a complete view over the database contents and avoid key collisions.

module Block_priority : sig ... end

module Roll : sig ... end

Storage from this submodule must only be accessed through the module `Roll`.

module Contract : sig ... end

Storage from this submodule must only be accessed through the module `Contract`.

module Big_map : sig ... end

module Delegates : Storage_sigs.Data_set_storage with type t := Raw_context.t and type elt = Tezos_protocol_environment.Signature.Public_key_hash.t

Set of all registered delegates.

module Active_delegates_with_rolls : Storage_sigs.Data_set_storage with type t := Raw_context.t and type elt = Tezos_protocol_environment.Signature.Public_key_hash.t

Set of all active delegates with rolls.

module Delegates_with_frozen_balance : Storage_sigs.Data_set_storage with type t = Raw_context.t * Cycle_repr.t and type elt = Tezos_protocol_environment.Signature.Public_key_hash.t

Set of all the delegates with frozen rewards/bonds/fees for a given cycle.

Votes

module Vote : sig ... end

Seed

module Seed : sig ... end

Storage from this submodule must only be accessed through the module `Seed`.

Commitments

module Commitments : Storage_sigs.Indexed_data_storage with type key = Blinded_public_key_hash.t and type value = Tez_repr.t and type t := Raw_context.t

Ramp up security deposits...

module Ramp_up : sig ... end